Generalise the Fast Stack API for cross-platform support.
 - Define a handler method which specifies the handler for the given platform, abstracting the handling of the timer signal away for platforms.
 - Rename fast_stack.c to fast_stack_linux.c
 - Build the Linux interface only if we have backtrace_symbols_fd.
