+++ ext/fast_stack/extconf.rb
@@ -5,4 +5,7 @@ require 'fileutils'
 # work around ruby 2.0 p0 bug
 FileUtils.mkdir_p(File.expand_path('../../../lib/fast_stack',__FILE__))
 
+have_func('backtrace_symbols_fd')
+
+
 create_makefile('fast_stack/fast_stack')
